chore(solana): fix warnings found by new IntelliJ (#8327)
bfb8abe3e9755b12eb9ed6dbd95dd91f543010a9

+++ src/main/kotlin/sbf/dwarf/DWARFEdgeLabelAnnotator.kt
@@ -33,7 +33,6 @@ import utils.*
 import vc.data.CoreTACProgram
 import vc.data.SnippetCmd
 import vc.data.TACMeta
-import vc.data.tacexprutil.*
 
 val logger = Logger(LoggerTypes.SBF_DEBUG_ANNOTATOR)
 
@@ -660,7 +659,7 @@ data class CallStackLevel private constructor(val scopes: Set<DwarfMethod>, val 
                     val registerAccess = ops.getRegisterAccess(currIns.access.base.r)
                     if (registerAccess != null) {
                         val registerOffset = registerAccess.offset()
-                        val offsetIntoStruct = currIns.access.offset - (frameBasePointer.toLong() + registerOffset)
+                        val offsetIntoStruct = currIns.access.offset - (frameBasePointer + registerOffset)
                         if (offset == Offset(offsetIntoStruct.toULong())) {
                             DWARFCfgEdgeLabel.DirectMemoryAccess(this, sourceVariable, offset, ops, currIns, currIns.value)
                         } else {
